server/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
idna==3.10
mysqlclient==2.2.5
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
